+++ pages/messages.tsx
@@ -1,87 +1,18 @@
 import { Snackbar } from "@material-ui/core"
 import { Alert } from "@material-ui/lab"
-import { makeStyles } from "@material-ui/styles"
-import React, { useEffect, useState } from "react"
-import { useDispatch, useSelector } from "react-redux"
-import actions from "src/actions"
-import ContentHeader from "src/components/design/ContentHeader"
-import FloatingDiv from "src/components/design/FloatingDiv"
-import Message from "src/components/design/messages/Message"
+import React, { useState } from "react"
+import { useSelector } from "react-redux"
 import ScottyLabsHeader from "src/components/design/ScottyLabsHeader"
 import WaveFooter from "src/components/design/WaveFooter"
 import Menu from "src/components/menu/Menu"
+import MessagesDialog from "src/components/teams/messages/MessagesDialog"
 import { AuthenticatedLayout } from "src/layouts"
 import { RootState } from "types/RootState"
 
-const useStyles = makeStyles((theme) => ({
-  tableData: {
-    tableLayout: "fixed",
-    width: "100%",
-    textAlign: "left",
-    borderCollapse: "collapse",
-    borderSpacing: "0 33px"
-  }
-}))
-
-interface RequestData {
-  seen: boolean
-  _id: string
-  type: string
-}
-
 const Messages = () => {
-  const dispatch = useDispatch()
   const errorMessage = useSelector((state: RootState) => state?.requests?.error)
-  const [requests, setRequests] = useState<any>([])
-  const [seen, setSeen] = useState<Array<boolean>>([])
   const [notify, setNotify] = useState("")
   const [successMessage, setSuccessMessage] = useState("")
-  const [isCaptain, setIsCaptain] = useState(false)
-  const classes = useStyles()
-  const user = useSelector((state: RootState) => state?.accounts?.data)
-
-  useEffect(() => {
-    if (user === undefined) {
-      return
-    }
-
-    const fetchRequests = async () => {
-      let captain = false
-      try {
-        const teamInfo = await dispatch(actions.user.getOwnTeam())
-        captain = user._id === teamInfo.data.admin
-      } catch (err) {
-        console.log(err)
-      }
-      try {
-        console.log(captain)
-        setIsCaptain(captain)
-        const req = captain
-          ? await dispatch(actions.requests.curTeamRequests())
-          : await dispatch(actions.requests.curUserRequests())
-        setRequests(req.data)
-        setSeen(
-          req.data.map((r: RequestData) => {
-            return r.seen
-          })
-        )
-        req.data.forEach(async (value: any) => {
-          if (value.type != "JOIN") {
-            await dispatch(actions.requests.openRequest(value._id))
-          }
-        })
-      } catch (err) {
-        console.log(err)
-      }
-    }
-
-    fetchRequests()
-  }, [user])
-
-  function handleRemove(id: number) {
-    const newRequests = requests.filter((item: any) => item.id !== id)
-    setRequests(newRequests)
-  }
 
   return (
     <>
@@ -89,24 +20,7 @@ const Messages = () => {
       <div>
         <ScottyLabsHeader />
         <WaveFooter />
-        <FloatingDiv>
-          <ContentHeader title="Messages" />
-          <table className={classes.tableData}>
-            <tbody>
-              {requests.map((request: any, idx: number) => (
-                <Message
-                  key={idx}
-                  content={request}
-                  setSuccessMessage={setSuccessMessage}
-                  setNotify={setNotify}
-                  isNew={!seen[idx]}
-                  isCaptain={isCaptain}
-                  handleRemove={handleRemove}
-                />
-              ))}
-            </tbody>
-          </table>
-        </FloatingDiv>
+        <MessagesDialog setSuccessMessage={setSuccessMessage} />
         <Snackbar
           open={notify != ""}
           autoHideDuration={5000}
